MNT: really rip get_datum out of retrive
 - create new module 'core' where DB manipulation logic will live
 - not intended for end-users, but part of public developer API
 - functions will be stateless and have very cranky/verbose signatures
 - use to back both FileStore objects for library use and server for
   service use.
